Add files via upload
This is mostly the summary of my work in The Tech Academy's SQL Server RDBMS class.  It consists of a single script DB setup with constraint and data population.  The second being the query drills that followed after its creation.  Will add the actual drill text to give some idea of the objective.